# docs/swagger/swagger.yml
swagger: '2.0'

info:
  version: "0.0.1"
  title: himawariAPI
basePath: /api/v1
produces:
  - application/json

paths:
  /timecards/{yearMonth}:
    get:
      summary: API0001:開始/終了時間取得
      description: |
        ログインユーザーに対応する対象月の日次の開始/終了日時の一覧を返します。
      tags:
        - timecard
      parameters:
        - name: yearMonth
          in: path
          description: 年月(yyyyMM)
          required: true
          type: string
      responses:
        200:
          description: Successful response
          schema:
            title: 対象月の日次の開始/終了日時
            type: object
            properties:
              yearMonth:
                title: 年月
                type: string
              days:
                $ref: '#/definitions/Days'
  /users/{userId}/timecards/{yearMonth}:
    get:
      summary: API1001:開始/終了時間取得
      description: |
        管理者権限のみアクセス可能です。
        対象月の日次の開始/終了日時の一覧を返します。
      tags:
        - timecard
      parameters:
        - name: userId
          in: path
          description: ユーザーID
          required: true
          type: string
        - name: yearMonth
          in: path
          description: 年月(yyyyMM)
          required: true
          type: string
      responses:
        200:
          description: Successful response
          schema:
            title: 対象月の日次の開始/終了日時
            type: object
            properties:
              yearMonth:
                title: 年月
                type: string
              days:
                $ref: '#/definitions/Days'
  /user/days:
    post:
      summary: API0002:日次情報登録
      description: 日単位の開始、終了時間を登録します。
      tags:
        - timecard
      parameters:
        - name: StartEndDatetimes
          in: body
          schema:
            title: 日次情報登録パラメータ
            type: object
            properties:
              days:
                $ref: '#/definitions/Days'
      responses:
        200:
          description: 登録結果
          schema:
            title: Result
            type: object
            properties:
              resultType:
                title: 登録結果コード
                type: string
                enum:
                  - success
                  - failure
              message:
                title: 結果メッセージ
                type: string
            required:
              - resultType

################################################################################
#                                 Definitions                                  #
################################################################################
definitions:
  Days:
    title: 日次情報リスト
    type: array
    items:
      title: Day:日次情報
      type: object
      properties:
        bizDate:
          title: 業務日
          description: yyyy-MM-dd
          type: string
        startDatetime:
          title: 開始日時
          description: yyyy-MM-ddTHH:mm:dd:ssZ
          type: string
        endDatetime:
          title: 終了日時
          description: yyyy-MM-ddTHH:mm:dd:ssZ
          type: string
        vacationTypeCode:
          $ref: '#/definitions/VacationType'
        note:
          title: 備考
          type: string
      required:
        - bizDate
  VacationType:
    title: 休暇タイプ
    type: string
    enum:
      - PAID_DAY_OFF
      - SP_DAY_OFF
      - AM_OFF
      - PM_OFF
      - TRANSFER_DAY_OFF
    description: |
      有給休暇
      特別休暇
      AM休
      PM休
      振替休暇